Bug-Fix: User Being Asked to Save Scan Settings Before All Prompts Are Answered
When a user runs our scan.py file, they are met with a series of interactive yes/no terminal prompts that allow them to customize their scan. Prior to this fix, the user was asked if they would like to save their scan settings for next time, before they had the opportunity to even answer all of the prompts. This has now been fixed with some additional improvements as well:

- The user's most recent "show collaboration info" choice is now displayed in the list of saved scan settings.
- Fixed some redundant code that caused the user to be prompted to "save their scan to the database file" multiple times in a row.
- Fixed an issue where the user was still being asked if they would like to "show collaboration info" after choosing to re-use the exact settings from their previous scan.