Bug-Fix: Unable to Overwrite a Scan's File Type Filtering Back to None
Prior to this fix, a user's local config.json file (which store previously-saved scan settings/parameters) had an issue with it's "file_type" field. By default it is set to None, and it was able to be overwritten to whatever saved file extension the user chose. However, once the file_type field had been changed from its default None value, it could not be changed back to None ever again due to logic issues associated with using update() on fields within config.json.

- config.py's save_config() and merge_settings() functions have been updated with improved, more robust logic that allows for the handling of None/empty values being written back into config.json's fields.
- config.py had its DEFAULT config.json values updated to include fields for data_consent (introduced with consent.py) and show_collaboration (introduced directly into scan.py)
- config_test.py has had its unit tests updated to test for the newly-introduced fields mentioned above. (In order for one of the tests to function properly, I had to update run_with_saved_settings() from scan.py as it had a logic issue in how it passed an args dictionary with empty keys, now it only passes keys into merge_settings() that were explicitly provided)