Updated the unit testing suite for data-access consent feature
Made a quick hotfix that fixes a unit test within consent_test.py that was using deprecated Strings from the previous implementation of our data-access policy. The test essentially ensures that when no strings are explicitly passed into describe_data_access(), it will return the default Strings that are hardcoded into the function. Now it checks that the four main headers from our updated data-access policy are found in the output. It is not entirely comprehensive, as I did not want to include 23 different self.assertIn() calls, but I think it is adequate for ensuring the default Strings are in the output. All 4 unit tests within consent_test.py are passing on my local machine as of Nov. 30, 2025 @ 11:40am.